新增基础控制器，支持token用户信息获取与重置，完善会议快速创建功能

+++ src/main/java/org/lzg/meeting/model/entity/User.java
@@ -80,7 +80,7 @@ public class User implements Serializable {
 	 * 会议号
 	 */
 	@TableField("meetingNo")
-	private Long meetingNo;
+	private Integer meetingNo;
 	/**
 	 * 创建时间
 	 */
